Rename dfttest file to fft
